+++ datumaro/plugins/coco_format/importer.py
@@ -21,6 +21,7 @@ class CocoImporter(Importer):
         CocoTask.captions: 'coco_captions',
         CocoTask.labels: 'coco_labels',
         CocoTask.image_info: 'coco_image_info',
+        CocoTask.panoptic: 'coco_panoptic',
     }
 
     @classmethod
@@ -39,7 +40,7 @@ class CocoImporter(Importer):
 
         # TODO: should be removed when proper label merging is implemented
         conflicting_types = {CocoTask.instances,
-            CocoTask.person_keypoints, CocoTask.labels}
+            CocoTask.person_keypoints, CocoTask.labels, CocoTask.panoptic}
         ann_types = set(t for s in subsets.values() for t in s) \
             & conflicting_types
         if 1 <= len(ann_types):